+++ levelup-react-16/src/App.js
@@ -2,13 +2,13 @@ import React from "react";
 import { BrowserRouter as Router, Route, Switch, Link } from "react-router-dom";
 import logo from "./logo.svg";
 import "./App.css";
-import MoviesList from "./MoviesList";
-import MoviesDetail from "./MovieDetail";
+import MoviesList from "./Movies/MoviesList";
+import MoviesDetail from "./Movies/MovieDetail";
 import { Provider } from "react-redux";
 import { applyMiddleware, createStore } from "redux";
 import rootReducer from "./rootReducer";
 import { composeWithDevTools } from "redux-devtools-extension";
-import Toggle from "./Toggle";
+import Toggle from "./Toggle/Toggle";
 import logger from "redux-logger";
 import thunk from "redux-thunk";
