+++ levelup-react-16/src/Movies/actions.js
@@ -1,11 +1,4 @@
-export const TOGGLE_MESSAGE = "TOGGLE_MESSAGE";
 export const GET_MOVIES = "GET_MOVIES";
-
-export function toggleMessage() {
-  return {
-    type: "TOGGLE_MESSAGE"
-  };
-}
 
 export function getMovies() {
   return async function(dispatch) {
