+++ levelup-react-16/src/Toggle/actions.js
@@ -1,0 +1,7 @@
+export const TOGGLE_MESSAGE = "TOGGLE_MESSAGE";
+
+export function toggleMessage() {
+  return {
+    type: "TOGGLE_MESSAGE"
+  };
+}
