+++ levelup-react-16/src/Toggle/reducer.js
@@ -1,18 +1,14 @@
-import { TOGGLE_MESSAGE, GET_MOVIES } from "./actions";
+import { TOGGLE_MESSAGE } from "./actions";
 
 const initialState = {
-  messageVisibility: false,
-  movies: []
+  messageVisibility: false
 };
 
 export default function(state = initialState, action) {
-  const { type, data } = action;
+  const { type } = action;
   switch (type) {
     case TOGGLE_MESSAGE:
       return { ...state, messageVisibility: !state.messageVisibility };
-
-    case GET_MOVIES:
-      return { ...state, movies: data };
 
     default:
       return state;
