+++ levelup-react-16/src/rootReducer.js
@@ -1,8 +1,10 @@
 import { combineReducers } from "redux";
-import message from "./reducer";
+import toggle from "./Toggle/reducer";
+import movies from "./Movies/reducer";
 
 const rootReducer = combineReducers({
-  message
+  toggle,
+  movies
 });
 
 export default rootReducer;
